[reader mode] Prevent unintended int conversion for score values
The scores are implicitly cast to ints which makes the values all be either 0 or 1. Not postfixing the histogram because these aren't really used right now. Instead adding a note that these are broken before M140.

Bug: 427690977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6672388
Commit-Queue: Brandon Wylie <wylieb@google.com>
Cr-Commit-Position: refs/heads/main@{#1478578}

+++ components/dom_distiller/content/renderer/distillability_agent.cc
@@ -119,8 +119,8 @@ void DumpDistillability(content::RenderFrame* render_frame,
                                     msg);
 }
 
-void RecordDistillabilityMetrics(int score,
-                                 int long_score,
+void RecordDistillabilityMetrics(double score,
+                                 double long_score,
                                  bool is_disitillable,
                                  ukm::UkmRecorder* ukm_recorder,
                                  ukm::SourceId source_id) {
